small cleanup / simplifications
