+++ aicssegmentation/workflow/__init__.py
@@ -1,5 +1,6 @@
 from .segmenter_function import SegmenterFunction, FunctionParameter, WidgetType  # noqa F401
 from .workflow_step import WorkflowStep, WorkflowStepCategory  # noqa F401
-from .workflow import Workflow, BatchWorkflow  # noqa F401
+from .workflow import Workflow  # noqa F401
+from .batch_workflow import BatchWorkflow  # noqa F401
 from .workflow_definition import WorkflowDefinition, PrebuiltWorkflowDefinition  # noqa F401
 from .workflow_engine import WorkflowEngine  # noqa F401
