+++ aicssegmentation/workflow/workflow.py
@@ -2,15 +2,11 @@ import numpy as np
 import logging
 
 from typing import Any, Dict, List
-from aicsimageio import AICSImage
-from aicsimageio.writers import OmeTiffWriter
+from aicssegmentation.exceptions import ArgumentNullError
 from .workflow_step import WorkflowStep
 from .workflow_definition import WorkflowDefinition
-from pathlib import Path
 
 log = logging.getLogger(__name__)
-
-SUPPORTED_FILE_EXTENSIONS = [".tiff", ".tif", ".czi"]
 
 
 class Workflow:
@@ -22,9 +18,9 @@ class Workflow:
 
     def __init__(self, workflow_definition: WorkflowDefinition, input_image: np.ndarray):
         if workflow_definition is None:
-            raise ValueError("workflow_definition")
+            raise ArgumentNullError("workflow_definition")
         if input_image is None:
-            raise ValueError("image")
+            raise ArgumentNullError("input_image")
         self._definition: WorkflowDefinition = workflow_definition
         self._starting_image: np.ndarray = input_image
         self._next_step: int = 0  # Next step to execute
@@ -167,164 +163,3 @@ class Workflow:
             (bool): True if all WorkflowSteps have been executed, False if not
         """
         return self._next_step >= len(self._definition.steps)
-
-
-class BatchWorkflow:
-    """
-    Represents a batch of workflows to process.
-    This class provides the functionality to run batches of workflows using multiple image inputs from a input directory
-    according to the steps defined in its WorkflowDefinition.
-    """
-
-    def __init__(
-        self, workflow_definition: WorkflowDefinition, input_dir: str, output_dir: str, channel_index: int = 0
-    ):
-        if workflow_definition is None:
-            raise ValueError("workflow_definition")
-        self._workflow_definition = workflow_definition
-
-        self.input_path = Path(input_dir)
-        if not self.input_path.exists():
-            raise ValueError("The input directory does not exist")
-
-        self.output_path = Path(output_dir)
-        # Creating an the output directory at output_dir if it does not exist already
-        if not self.output_path.exists():
-            if self.output_path.parent.exists():
-                self.output_path.mkdir(parents=True, exist_ok=True)
-            else:
-                raise ValueError(f"Output directory does not exist, and cannot be created at {self.output_path.parent}")
-
-        self._files_count: int = 0
-        self._failed_files: int = 0
-        self._channel_index = channel_index
-        self._log_file: Path = self.output_path.joinpath("log.txt")
-        self.write_to_log_file("Log for batch processing run\n", True)
-
-    def is_valid_image(self, image_path: Path) -> bool:
-        """
-        Check if file at a given image_path and is a valid image type we support.
-
-        Params:
-            image_path (Path): image to check
-
-        Returns:
-            (bool): True if file has a supported file extension.
-        """
-        if not image_path.exists():
-            return False
-        if image_path.suffix.lower() in SUPPORTED_FILE_EXTENSIONS:
-            return True
-        else:
-            return False
-
-    def process_all(self):
-        """
-        Process all images in the input_dir with the workflow_definition used to set up the BatchWorkflow
-
-        Params:
-            none
-
-        Returns:
-            none
-        """
-
-        files = [f for f in self.input_path.glob("**/*") if f.is_file]
-        # Currently will save files in same format as they are in the input path
-        for f in files:
-            self._files_count += 1
-            if self.is_valid_image(f):
-                read_image = AICSImage(f)
-
-                try:
-                    # read and format image in the way we expect
-                    image_from_path = self.format_image_to_3d(read_image)
-                    # Run workflow on image
-                    workflow = Workflow(self._workflow_definition, image_from_path)
-                    result = workflow.execute_all()
-                    with OmeTiffWriter(self.output_path.joinpath(f.name), overwrite_file=True) as w:
-                        w.save(data=self.convert_bool_to_uint8(result), dimension_order="ZYX")
-
-                except Exception as e:
-                    # Handle failures during workflow execution/save
-                    self._failed_files += 1
-                    self.write_to_log_file(f"FAILED: {f}, ERROR: {e}\n")
-            else:
-                self._failed_files += 1
-                self.write_to_log_file(f"FAILED: {f}, ERROR: Unsupported Image Type {f.suffix}\n")
-        self._write_log_file_summary()
-
-    def _write_log_file_summary(self):
-        """
-        Write a log file to the output folder.
-
-        Params:
-            none
-
-        Returns:
-            none
-        """
-        if self._files_count == 0:
-            report = "There were no files to process in the input directory"
-        else:
-
-            files_processed = self._files_count - self._failed_files
-            report = f"{files_processed}/{self._files_count} files were processed.\n"
-        self.write_to_log_file(report + "\n")
-
-    def format_image_to_3d(self, image: AICSImage) -> np.ndarray:
-        """
-        Format images in the way that aics-segmention expects for most workflows (3d, zyx)
-
-        Params:
-            image_path (AICSImage): image to format
-
-        Returns:
-            np.ndarray: segment-able image for aics-segmentation
-        """
-        dims = len(image.shape) - image.shape.count(1)
-        if dims == 6:
-            # STCZYX
-            return TypeError("6D images are unsupported.")
-        elif dims == 5:
-            return TypeError("5D images are unsupported.")
-            # if "S" in image.dims.order and "C" in image.dims.order:
-            #     return image.get_image_data("ZYX", C=self._channel_index, S=0)
-            # elif "T" in image.dims.order and "C" in image.dims.order:
-            #     return image.get_image_data("ZYX", C=self._channel_index, T=0)
-            # elif "S" in image.dims.order and "T":
-            #     return image.get_image_data("ZYX", S=0, T=0)
-        elif dims == 4:
-            if "S" in image.dims.order:
-                return TypeError("Multi-Scene images are unsupported")
-            elif "T" in image.dims.order:
-                return TypeError("Timelapse images are unsupported.")
-            else:
-                return image.get_image_data("ZYX", C=self._channel_index)
-        elif dims == 3:
-            return image.get_image_data("ZYX")
-        else:
-            return TypeError(f"Unsupported image format {image.dims.order}")
-
-    def convert_bool_to_uint8(self, image: np.ndarray):
-        """
-        Format segmented images to uint8 to save via AICSImage
-
-        Params:
-            image (np.ndarray): segmented image
-
-        Returns:
-            np.ndarray: image converted to uint8 for saving
-        """
-        image = image.astype(np.uint8)
-        image[image > 0] = 255
-        return image
-
-    def write_to_log_file(self, write: str, new_file: bool = False):
-        if new_file:
-            with open(self._log_file, "w") as writer:
-                writer.write(write)
-        else:
-            # Append to old file
-            with open(self._log_file, "a") as writer:
-                writer.write(write)
