+++ aicssegmentation/workflow/workflow_engine.py
@@ -1,7 +1,9 @@
 import numpy as np
 
 from typing import List, Union
-from .workflow import Workflow, BatchWorkflow
+from aicssegmentation.exceptions import ArgumentNullError
+from .workflow import Workflow
+from .batch_workflow import BatchWorkflow
 from .workflow_definition import WorkflowDefinition
 from .workflow_config import WorkflowConfig
 from pathlib import Path
@@ -32,8 +34,10 @@ class WorkflowEngine:
             workflow_name (str): Name of the workflow to load
             input_image (ndarray): input image for the workflow to execute on
         """
+        if workflow_name is None:
+            raise ArgumentNullError("workflow_name")
         if input_image is None:
-            raise ValueError("input_image is None")
+            raise ArgumentNullError("input_image")
 
         definition = self._get_workflow_definition(workflow_name)
 
@@ -51,6 +55,13 @@ class WorkflowEngine:
             output_dir (str|Path): Output directory for the batch processing
             channel_index (int): Index of the channel to process in each image (usually a structure channel)
         """
+        if workflow_name is None:
+            raise ArgumentNullError("workflow_name")
+        if input_dir is None:
+            raise ArgumentNullError("input_dir")
+        if output_dir is None:
+            raise ArgumentNullError("output_dir")
+
         definition = self._get_workflow_definition(workflow_name)
 
         return BatchWorkflow(definition, input_dir, output_dir, channel_index)
@@ -66,9 +77,9 @@ class WorkflowEngine:
             input_image (ndarray): input image for the workflow to execute on
         """
         if input_image is None:
-            raise ValueError("input_image is None")
+            raise ArgumentNullError("input_image")
         if file_path is None:
-            raise ValueError("file_path is None")
+            raise ArgumentNullError("file_path")
 
         definition = self._workflow_config.get_workflow_definition_from_config_file(Path(file_path))
         return Workflow(definition, input_image)
@@ -89,17 +100,21 @@ class WorkflowEngine:
             output_dir (str|Path): Output directory for the batch processing
             channel_index (int): Index of the channel to process in each image (usually a structure channel)
         """
-        # TODO more input validations
-        # TODO it's time for a ArgumentNullError Exception type
+        if file_path is None:
+            raise ArgumentNullError("file_path")
+        if input_dir is None:
+            raise ArgumentNullError("input_dir")
+        if output_dir is None:
+            raise ArgumentNullError("output_dir")
 
         definition = self._workflow_config.get_workflow_definition_from_config_file(Path(file_path))
         return BatchWorkflow(definition, input_dir, output_dir, channel_index)
 
     def save_workflow_definition(self, workflow_definition: WorkflowDefinition, output_file_path: Union[str, Path]):
         if workflow_definition is None:
-            raise ValueError("workflow_definition is None")
+            raise ArgumentNullError("workflow_definition")
         if output_file_path is None:
-            raise ValueError("file_path is None")
+            raise ArgumentNullError("file_path")
 
         self._workflow_config.save_workflow_definition_as_json(workflow_definition, output_file_path)
